backend-py/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Joey Xiong <junruxiong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
deepgram-sdk = "^3.3.1"
python-dotenv = "^1.0.1"
pyaudio = "^0.2.14"
aiohttp = "^3.9.5"
websockets = "^12.0"
pydub = "^0.25.1"
pydantic-settings = "^2.3.4"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
httpx = "^0.27.0"
ffmpeg = "^1.4"
celery = "^5.4.0"
redis = "^5.0.7"
flower = "^2.0.1"
python-jose = "^3.3.0"
supabase = "^2.5.1"
anthropic = ">=0.28.0,<0.29.0"
pytest = "^8.2.2"
transformers = "4.38.2"
sounddevice = "^0.4.7"
rhasspy-silence = "^0.4.0"
pygame = "^2.6.0"
speechrecognition = "^3.10.4"
onnxruntime = "^1.18.1"
torch = "2.3.0"
torchaudio = "2.3.0"
colorama = "^0.4.6"
azure-cognitiveservices-speech = "^1.38.0"
emoji = "^2.12.1"
funasr = "^1.1.3"
litellm = "^1.43.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"